feat: clipboard docs
