+++ src/lib/index.ts
@@ -11,6 +11,7 @@ export * from "./buttons";
 export * from "./card";
 export * from "./carousel";
 export * from "./chart";
+export * from "./clipboard";
 export * from "./darkmode";
 export * from "./datepicker";
 export * from "./device-mockups";
